Rust - We're back with ownership
We're back…
We're back!
In our own backyard!
We're gonna make the most of every day.
We've got plans.
They may be avant-garde,
But ya know we're gonna do them anyway.

It feels so good to be where we belong!
I guess that's the main bullet point of this little song
We're tuned up and right on track.
Baby, we're back!
Baby, we're back!

I'm back!
She's back!
And I'll tell you why:
There's still some things to be discussed.
Like how every day I try and try
And yet here we are and I still got some brothers to bust!

I work so hard. I plan and follow through,
But my mom still never sees anything that they do!
This time, I'm gonna launch a counterattack.
Baby, I'm back!
Baby, we're back!

Yes, all our friends are here!
And Perry still tends to disappear.
But we're havin' fun and livin' large.
And this time, I'm gonna be in charge. (Be in charge… be in charge…)

Well, you know. Conditionally.

I'm back! I don't wanna scare ya,
And it may sound superfluous,
But I can really take over the Tri-State Area
If I can just defeat this platypus.

We're back! Yeah, it's a brand new day.
We're rested and we're ready to jump into the fray.
So you can put that in your almanac.
Baby, we're back!
Baby, we're back!

We're back! And we're feelin' groovy.
On Disney+, you can watch our brand new movie!
It's just a simple statement of facts.
Baby, we're back!
Baby, we're back!
Baby, we're back!
Baby, we're back!
Baby, we're back!
Baby, we're back!
Baby, we're back!